+++ src/main/java/it/unicam/cs/mpgc/rpg125936/domain/shop/ToolOffer.java
@@ -5,10 +5,10 @@ import it.unicam.cs.mpgc.rpg125936.domain.item.Pickaxe;
 public class ToolOffer {
     private final long id;
     private final String name;
-    private final double maxUses;
+    private final int maxUses;
     private final double price;
 
-    public ToolOffer(long id, String name, double maxUses, double price) {
+    public ToolOffer(long id, String name, int maxUses, double price) {
         this.id = id;
         this.name = name;
         this.maxUses = maxUses;
@@ -17,7 +17,7 @@ public class ToolOffer {
 
     public long getId() { return id; }
     public String getName() { return name; }
-    public double getMaxUses() { return maxUses; }
+    public int getMaxUses() { return maxUses; }
     public double getPrice() { return price; }
 
     public Pickaxe createItem() {
