+++ src/main/java/it/unicam/cs/mpgc/rpg125936/service/mine/MiningResultDTO.java
@@ -6,9 +6,9 @@ public class MiningResultDTO {
     private final boolean success;
     private final String message;
     private final Material material;
-    private final double remainingDurability;
+    private final int remainingDurability;
 
-    public MiningResultDTO(boolean success, String message, Material material, double remainingDurability) {
+    public MiningResultDTO(boolean success, String message, Material material, int remainingDurability) {
         this.success = success;
         this.message = message;
         this.material = material;
@@ -18,5 +18,5 @@ public class MiningResultDTO {
     public boolean isSuccess() { return success; }
     public String getMessage() { return message; }
     public Material getMaterial() { return material; }
-    public double getRemainingDurability() { return remainingDurability; }
+    public int getRemainingDurability() { return remainingDurability; }
 }
